+++ R/omop_join_name.R
@@ -6,6 +6,9 @@
 #' @param df dataframe
 #' @param namestart start of *_concept_id column, if "" will join on concept_name, ignored if namefull used
 #' @param namefull optional full name of concept_id column, if default "" namestart used
+#' @param domain whether to add domain_id column, default FALSE
+#' @param vocabulary whether to add vocabulary_id column, default FALSE
+#' @param concept_class whether to add concept_class_id column, default FALSE
 #' @export
 #' @examples
 #' data.frame(concept_id=(c(3571338L,4002075L))) |> omop_join_name()
@@ -16,7 +19,12 @@
 #' #       dplyr::collect() |> omop_join_name(namefull="concept_id_1")
 omop_join_name <- function(df,
                            namestart = "",
-                           namefull = "") {
+                           namefull = "",
+                           domain = FALSE,
+                           vocabulary = FALSE,
+                           concept_class = FALSE
+                           #todo can add standard_concept & concept_code
+                           ) {
 
   #"" is to cope with concept_id from omop_join_name_all()
   if (namefull != "") id_col_name <- namefull
@@ -30,9 +38,22 @@ omop_join_name <- function(df,
 
   #beware rename concept_name column before joining in case
   #there is already a concept_name column in df
+
+  columns2join <- c("concept_id","concept_name")
+
+  #2023-10-05 optionally include vocabulary_id and domain_id
+  #protect against problem if more than 1 id column it can become .x,.y etc.
+  if (domain == TRUE & !("domain_id" %in% names(df)))
+    columns2join <- c(columns2join, "domain_id")
+  if (vocabulary == TRUE & !("vocabulary_id" %in% names(df)))
+    columns2join <- c(columns2join, "vocabulary_id")
+  if (concept_class == TRUE & !("concept_class_id" %in% names(df)))
+    columns2join <- c(columns2join, "concept_class_id")
+
   id_and_name <- omopcept::omop_concept() |>
-    select(.data$concept_id,.data$concept_name) |>
-    rename_with(~name_col_name, .data$concept_name)
+    select(any_of(columns2join)) |>
+    #change concept_name to the full name
+    rename_with(~name_col_name, concept_name)
 
   #TODO can I make this faster by replacing the copy=TRUE
   #with some filter & collect ?
@@ -61,6 +82,10 @@ ojoin <- omop_join_name
 #' e.g. drug_concept_id to get drug_name etc.
 #'
 #' @param df dataframe
+#' @param domain whether to add domain_id column, default FALSE
+#' @param vocabulary whether to add vocabulary_id column, default FALSE
+#' @param concept_class whether to add concept_class_id column, default FALSE
+#'
 #' @return dataframe based on input df with 1 extra column added for each concept_id column
 #' @export
 #' @examples
@@ -76,7 +101,10 @@ ojoin <- omop_join_name
 #' #       omop_join_name_all()
 #' #df3 <- omop_concept_relationship() |> head() |>
 #' #          dplyr::collect() |> omop_join_name_all()
-omop_join_name_all <- function(df) {
+omop_join_name_all <- function(df,
+                               domain = FALSE,
+                               vocabulary = FALSE,
+                               concept_class = FALSE) {
 
   #logic
   #if colname contains *_concept_id do omop_join_name(namestart=*)
@@ -91,9 +119,9 @@ omop_join_name_all <- function(df) {
   for(cname in colnames)
   {
     if (str_detect(cname,"concept_id")) {
-          df <- df |> omop_join_name(namefull = cname)
+          df <- df |> omop_join_name(namefull=cname, domain=domain, vocabulary=vocabulary, concept_class=concept_class )
     } else
-          df <- df |> omop_join_name(namestart = cname)
+          df <- df |> omop_join_name(namestart=cname, domain=domain, vocabulary=vocabulary, concept_class=concept_class )
   }
 
   return(df)